% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_celltypes.R
\name{anno_celltypes}
\alias{anno_celltypes}
\title{Annotate each cell with the corresponding cell type}
\usage{
anno_celltypes(object, anno_level = 2, species = "Hs", ...)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A Seurat object
}
\description{
Annotate each cell with the corresponding cell type
}
\examples{
\dontrun{
sobj <- anno_celltypes(object = sobj, anno_level = 2, species = "Hs")
}
}
