% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_celltypes.R
\name{cluster_data}
\alias{cluster_data}
\title{Title}
\usage{
cluster_data(object, resolution = 0.8, npcs = NULL, seed = 42)
}
\arguments{
\item{object}{A}

\item{resolution}{A}
}
\value{
object
}
\description{
Title
}
\examples{
\dontrun{
cluster_data(object, resolution = 0.8)
}
}
