% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{create_adt_predictor}
\alias{create_adt_predictor}
\title{Create a Gene expression to ADT assay predictor}
\usage{
create_adt_predictor(do_log1p = FALSE)
}
\arguments{
\item{do_log1p}{}
}
\value{
pipe an adt predictor object
}
\description{
Create a Gene expression to ADT assay predictor
}
\examples{
\dontrun{
create_adt_predictor(do_log1p = FALSE)
}
}
