% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{evaluate_predictor}
\alias{evaluate_predictor}
\title{Evaluate the adt predictor}
\usage{
evaluate_predictor(
  pipe,
  gexp_test,
  adt_test,
  slot = "counts",
  normalize_gex = TRUE,
  normalize_adt = TRUE
)
}
\arguments{
\item{pipe}{Trained ADT predictor}

\item{gexp_test}{Matrix with gene expression data.}

\item{adt_test}{Matrix with ADT count data.}

\item{do_log1p}{A}
}
\value{
Returns a data frame containing RSME, Pearson correlation and Spearman correlation of the tested data.
}
\description{
Evaluate the adt predictor
}
\examples{
\dontrun{
evaluate_predictor(pipe, gex_test, adt_test)
}
}
