% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{gexp_normalize}
\alias{gexp_normalize}
\title{Normalize gene expression matrix with scran and scuttle}
\usage{
gexp_normalize(gexp_matrix, center.size.factors = FALSE, log = FALSE, ...)
}
\arguments{
\item{gexp_matrix}{A gene expression matrix}

\item{center.size.factors}{A}

\item{log}{A}

\item{...}{For the method, additional arguments passed to logNormCounts.}
}
\value{
Normalized expression matrix
}
\description{
Normalize gene expression matrix with scran and scuttle
}
\examples{
\dontrun{
# Normalize expression matirx
normalized_matrix <- gexp_normalize(sobj\@assays[["RNA"]]\@counts)
# Add normalized matrix back to RNA assay in Seurat.
sobj\@assays[["RNA"]]\@data <- normalized_matrix
}
}
