% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_pyton_dependencies}
\alias{install_pyton_dependencies}
\title{Install all python dependencies for scLinear}
\usage{
install_pyton_dependencies()
}
\description{
Install all python dependencies for scLinear
}
\examples{
\dontrun{
install_pyton_dependencies()
}
}
