% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_celltypes.R
\name{integrate_samples}
\alias{integrate_samples}
\title{Title}
\usage{
integrate_samples(object, method = "rpca", samples = "samples")
}
\arguments{
\item{object}{A}

\item{method}{A}

\item{samples}{A}

\item{resolution}{A}
}
\value{
object
}
\description{
Title
}
\examples{
\dontrun{
#' integrate_samples()
}
}
