% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{load_pretrained_model}
\alias{load_pretrained_model}
\title{Load a pretrained model}
\usage{
load_pretrained_model(pipe, model = "all")
}
\arguments{
\item{pipe}{A pipe}

\item{model}{Choose pretrained model: all, bcells, tcells, nkcells}
}
\value{
pipe
}
\description{
Load a pretrained model
}
\examples{
\dontrun{
load_pretrained_model(pipe, model = "all")
}
}
