% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{mad_filtering}
\alias{mad_filtering}
\title{Title}
\usage{
mad_filtering(object = objec, nmads = 3, type = "both", mttype = "lower")
}
\arguments{
\item{object}{Seurat object}

\item{nmads}{Median absolute deviation that should be counted as outliner}

\item{type}{If upper, lower , both boundaries should be used for filtering}

\item{mttype}{If upper, lower , both boundaries should be used for mitochondrial percent filtering}
}
\value{
object Seurat object
}
\description{
Title
}
\examples{
sobj <- mad_filtering(sobj)
}
