% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for modality prediction}
\usage{
prepare_data(
  object,
  remove_doublets = TRUE,
  low_qc_cell_removal = TRUE,
  anno_level = 2,
  samples = NULL,
  integrate_data = FALSE,
  remove_empty_droplets = FALSE,
  lower = 100,
  FDR = 0.01,
  annotation_selfCluster = FALSE,
  resolution = 0.8,
  seed = 42,
  return_plots = FALSE,
  print_plots = TRUE,
  species = "Hs",
  min.features = NULL,
  verbose = FALSE
)
}
\arguments{
\item{object}{A Seurat object}

\item{remove_doublets}{Should doublets be removed? (TRUE,FALSE)}

\item{low_qc_cell_removal}{Should low quality cells be removed using median absolute deviation?}

\item{anno_level}{Level of annotation used for cell type annotation. Either a single number or a vector with multiple numbers c(1,2,3,4).}

\item{samples}{Variable that contains the sample names. If set, will use rpca to integrate the samples.}

\item{integrate_data}{Should the sample be integrated through the rpca method?}

\item{remove_empty_droplets}{If the raw unfiltered matrix was used to create the Seurat object, empty droplets should be filtered out. (TRUE, FALSE)}

\item{lower}{Lower boundary to define empty droplets. All cells with a lower amount of nFeatures are assumed to be empty.}

\item{FDR}{FDR threshold to define non empty cells.}

\item{annotation_selfCluster}{Should the clusters determined by Seurat be used for cell type annotation.}

\item{resolution}{Resolution for louvain clustering.}

\item{seed}{Used seed.}

\item{return_plots}{Should plots be returned from function}

\item{print_plots}{Print plots. (TRUE, FALSE)}

\item{species}{Species. Relevant for cell type annotation. ("Hs", "Mm")}

\item{min.features}{Minimum ammount of features per cell. Replaces automatically determined threshold if bigger.}
}
\value{
object A pre-processed Seurat object  with annotated cell types
}
\description{
Prepare data for modality prediction
}
\examples{
\dontrun{
sobj <- scLinear(object = sobj, remove_doublets = TRUE, low_qc_cell_removal = TRUE, anno_level = 2, samples = NULL, integrate_data = FALSE, resolution = 0.8)
}
}
