% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess single cell sequencing data}
\usage{
preprocess_data(object = object)
}
\arguments{
\item{object}{A seurat object}
}
\value{
A seurat object with preprocessed data
}
\description{
Preprocess single cell sequencing data
}
\examples{
preprocess_data(object = sobj)
}
