% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{remove_doublets}
\alias{remove_doublets}
\title{Remove Heterotypic doublets from sequencing data}
\usage{
remove_doublets(
  object = object,
  samples = NULL,
  remove_cells = TRUE,
  seed = 42,
  print_plots = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{samples}{meta.data information with  the sequencing samples}

\item{remove_cells}{should the low quality cell be removed before returning the object}

\item{seed}{Seed that should be used}
}
\value{
list containing Seurat object and quality plots
}
\description{
Remove Heterotypic doublets from sequencing data
}
\examples{
\dontrun{
remove_doublets(object = object, samples = "batch")
}
}
