% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{save_trained_model}
\alias{save_trained_model}
\title{Save a trained model}
\usage{
save_trained_model(pipe, file = NULL)
}
\arguments{
\item{pipe}{A ADT predictor object}

\item{file}{name to save the model}
}
\value{
pipe Returns the pipe with a loaded pre-trained model.
}
\description{
Save a trained model
}
\examples{
\dontrun{
save_trained_model(pipe, file = "trained_pipe.joblib")
}
}
