% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{scLinear}
\alias{scLinear}
\title{Predict modalities based on gene expression data}
\usage{
scLinear(object = object, cell_type)
}
\arguments{
\item{object}{}
}
\value{
object A Seurat object containing additional single cell modalities
}
\description{
Predict modalities based on gene expression data
}
\examples{
\dontrun{
sobj <- scLinear(object = sobj)
}
}
