% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_celltypes.R
\name{visualize_data}
\alias{visualize_data}
\title{Create A UMAP from the annotated Seurat object}
\usage{
visualize_data(object, group.by = "cell_type", ndims = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{Paraneter by witch to group visualization by}

\item{ndims}{Number of dimensions for dimensional reduction}
}
\value{
p A Dim plot
}
\description{
Create A UMAP from the annotated Seurat object
}
\examples{
\dontrun{
p <- visualize_data(sobj, group.by = "cell_type", ndims = 30)
}
}
