% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{adt_predict}
\alias{adt_predict}
\title{Predict ADT values from gene expression}
\usage{
adt_predict(pipe, gexp)
}
\arguments{
\item{gexp}{}
}
\value{
adt_assay retuns an adt assay object
}
\description{
Predict ADT values from gene expression
}
\examples{
\dontrun{
adt_predict(gextp)
}
}
