% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{empty_drops}
\alias{empty_drops}
\title{Remove empty droplets}
\usage{
empty_drops(object, lower = 100, FDR = 0.01, samples = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{lower}{A}

\item{FDR}{A}
}
\value{
object A Seurat object
}
\description{
Remove empty droplets
}
\examples{
\dontrun{
empty_drops(object = object, lower = 100, FDR = 0.01)
}
}
