% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{evaluate_predictor}
\alias{evaluate_predictor}
\title{Evaluate the adt predictor}
\usage{
evaluate_predictor(pipe, gexp_test, adt_test, do_log1p = TRUE)
}
\arguments{
\item{pipe}{A}

\item{gexp_test}{A}

\item{adt_test}{A}

\item{do_log1p}{A}
}
\value{
A
}
\description{
Evaluate the adt predictor
}
\examples{
\dontrun{
evaluate_predictor(pipe, gex_test, adt_test)
}
}
