% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{fit_predictor}
\alias{fit_predictor}
\title{Train a predictor object}
\usage{
fit_predictor(
  pipe,
  gexp_train,
  adt_train,
  gexp_test = NULL,
  slot_gex = "counts",
  slot_adt = "counts",
  normalize_gex = TRUE,
  normalize_adt = TRUE,
  margin = 2
)
}
\arguments{
\item{pipe}{A predictor object}

\item{adt_train}{ADT assay}

\item{margin}{Same as margin in NormalizeData function from Seurat}

\item{gex_train}{Gene expression assay}

\item{normalize}{Normalize GEX and ATD assay before fitting.}
}
\value{
pipe A trained predictor object
}
\description{
Train a predictor object
}
\examples{
\dontrun{
fit_predictor(pipe = pipe, gex_train = object@assays$RNA , adt_train = object@assays$ADT)
}
}
