% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{fit_predictor}
\alias{fit_predictor}
\title{Train a predictor object}
\usage{
fit_predictor(pipe, gex_train, adt_train)
}
\arguments{
\item{pipe}{A predictor object}

\item{gex_train}{Gene expression assay}

\item{adt_train}{ADT assay}
}
\value{
pipe A trained predictor object
}
\description{
Train a predictor object
}
\examples{
\dontrun{
fit_predictor(pipe = pipe, = object@assays$RNA , adt_train = object@assays$ADT)
}
}
