% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{load_pretrained_model}
\alias{load_pretrained_model}
\title{Load a pre-trained model}
\usage{
load_pretrained_model(pipe, model = "all")
}
\arguments{
\item{pipe}{A pipe}

\item{model}{Choose a pre-trained model to load. The pretraines models were
trained on the NeurIPS data. In most cases we would recommend to use the model ("all") trained on all
cell types. models available: all, bcells, tcells, nkcells.}
}
\value{
pipe Returns the pipe with a loaded pre-trained model.
}
\description{
Load a pre-trained model
}
\examples{
\dontrun{
load_pretrained_model(pipe, model = "all")
}
}
