% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{mad_filtering}
\alias{mad_filtering}
\title{Remove low quality cells based on the median absolute deviation.}
\usage{
mad_filtering(
  object = objec,
  samples = NULL,
  nmads = 3,
  type = "both",
  mttype = "higher",
  remove_cells = TRUE,
  print_plots = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{nmads}{Median absolute deviation that should be counted as outlier}

\item{type}{If upper, lower , both boundaries should be used for filtering (for nCount_RNA and nFeature_RNA)}

\item{mttype}{If upper, lower , both boundaries should be used for mitochondrial percent filtering}

\item{remove_cells}{should the low quality cell be removed before returning the object}
}
\value{
list containint a Seurat object and quality plots
}
\description{
Remove low quality cells based on the median absolute deviation.
}
\examples{
\dontrun{
sobj <- mad_filtering(sobj)
}
}
