% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for modality prediction}
\usage{
prepare_data(
  object,
  remove_doublets = TRUE,
  low_qc_cell_removal = TRUE,
  anno_level = 2,
  samples = NULL,
  integrate_data = FALSE,
  remove_empty_droplets = FALSE,
  lower = 100,
  FDR = 0.01,
  annotation_selfCluster = FALSE,
  resolution = 0.8,
  seed = 42,
  return_plots = FALSE,
  print_plots = TRUE
)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
object A pre-processed Seurat object  with annotated cell types
}
\description{
Prepare data for modality prediction
}
\examples{
\dontrun{
sobj <- scLinear(object = sobj, remove_doublets = TRUE, low_qc_cell_removal = TRUE, anno_level = 2, samples = NULL, integrate_data = FALSE, resolution = 0.8)
}
}
