% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for modality prediction}
\usage{
prepare_data(
  object = object,
  remove_doublets = TRUE,
  low_qc_cell_removal = TRUE,
  anno_level = 2,
  samples = samples
)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
object A Seurat object containing with annotated cell types
}
\description{
Prepare data for modality prediction
}
\examples{
\dontrun{
sobj <- scLinear(object = sobj)
}
}
