% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{remove_doublets}
\alias{remove_doublets}
\title{Remove Heterotypic doublets from sequencing data}
\usage{
remove_doublets(object = object, samples = "samples")
}
\arguments{
\item{object}{A Seurat object}

\item{samples}{meta.data information with  the sequencing samples}
}
\value{
object A Seurat object with removed doublets
}
\description{
Remove Heterotypic doublets from sequencing data
}
\examples{
remove_doublets(object = object, samples = "batch")
}
