% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{scLinear}
\alias{scLinear}
\title{Predict modalities based on gene expression data}
\usage{
scLinear(object = object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
A Seurat object containing additional single cell modalities
}
\description{
Predict modalities based on gene expression data
}
\examples{
sobj <- scLinear(object = sobj)
}
