% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{scLinear}
\alias{scLinear}
\title{Predict modalities based on gene expression data}
\usage{
scLinear(
  object,
  remove_doublets = TRUE,
  low_qc_cell_removal = TRUE,
  anno_level = 2,
  samples = NULL,
  integrate_data = FALSE,
  remove_empty_droplets = FALSE,
  lower = 100,
  FDR = 0.01,
  annotation_selfCluster = FALSE,
  resolution = 0.8,
  seed = 42,
  return_plots = FALSE,
  model = "all",
  assay_name = "RNA",
  print_plots = FALSE,
  species = "Hs",
  min.features = NULL
)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
object A Seurat object containing additional single cell modalities
}
\description{
Predict modalities based on gene expression data
}
\examples{
\dontrun{
sobj <- scLinear(object = sobj)
}
}
